requirements.txt

flask==2.3.3
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
textblob==0.17.1
requests==2.31.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.7
transformers==4.33.2
torch==2.0.1